refactor: move ape-test cfg to config module
