+++ src/ape_test/provider.py
@@ -17,7 +17,6 @@ from web3.exceptions import ContractLogicError as Web3ContractLogicError
 from web3.providers.eth_tester.defaults import API_ENDPOINTS, static_return
 from web3.types import TxParams
 
-from ape.api.config import PluginConfig
 from ape.api.providers import BlockAPI, TestProviderAPI
 from ape.api.trace import TraceAPI
 from ape.api.transactions import ReceiptAPI, TransactionAPI
@@ -35,17 +34,13 @@ from ape.types.address import AddressType
 from ape.types.events import ContractLog, LogFilter
 from ape.types.vm import BlockID, SnapshotID
 from ape.utils.misc import gas_estimation_error_message
-from ape.utils.testing import DEFAULT_TEST_CHAIN_ID, DEFAULT_TEST_HD_PATH
+from ape.utils.testing import DEFAULT_TEST_HD_PATH
 from ape_ethereum.provider import Web3Provider
 from ape_ethereum.trace import TraceApproach, TransactionTrace
+from ape_test.config import EthTesterProviderConfig
 
 if TYPE_CHECKING:
     from ape.api.accounts import TestAccountAPI
-
-
-class EthTesterProviderConfig(PluginConfig):
-    chain_id: int = DEFAULT_TEST_CHAIN_ID
-    auto_mine: bool = True
 
 
 class LocalProvider(TestProviderAPI, Web3Provider):
